feat: integrate A10 and A11 roadmap tasks and introduce arandu_base
- Introduce arandu_base crate containing unified core utility modules: span, line_index, source_registry, stable_id, index_vec, string_pool, arena, vm, and bitset.
- Redefine Span and Token to be packed 12-byte representations for improved memory performance and cache locality.
- Implement type-safe GenerationalId, O(1) SlotMap/DenseSlotMap, and collision-detecting StableHandle.
- Migrate compiler subsystems (lexer, parser, semantics, diagnostics, CLI) and tests to use offset-based Spans and SourceRegistry.
- Update semantic tests golden files.

+++ crates/arandu_semantics/src/resolved.rs
@@ -9,8 +9,8 @@ pub type DocCommentMap = FxHashMap<NodeKey, Vec<String>>;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct NodeKey {
-    pub start: usize,
-    pub end: usize,
+    pub start: u32,
+    pub end: u32,
 }
 
 impl From<Span> for NodeKey {
